requirements.txt
fastapi==0.112.2
uvicorn==0.30.5
pandas==2.1.4
numpy==1.24.3
setuptools>=65.0.0
wheel>=0.40.0
python-dotenv==1.0.1
requests==2.32.3
